align client api function names and behavior

+++ test/page_test.c
@@ -88,7 +88,7 @@ static void page_reopen_test(void)
     rc = page_init(&page, "/tmp/store/page.resql", -1, 0);
     assert(rc == RS_OK);
 
-    for (int i = 0; i < 1000; i++) {
+    for (uint64_t i = 0; i < 1000; i++) {
         entry = page_entry_at(&page, prev_index + 1 + i);
         assert(entry != NULL);
         assert(entry_term(entry) == i);
@@ -126,7 +126,7 @@ static void page_expand_test(void)
 
     assert(page_expand(&page) == RS_OK);
 
-    for (int i = 0; i < 1000; i++) {
+    for (uint64_t i = 0; i < 1000; i++) {
         entry = page_entry_at(&page, prev_index + 1 + i);
         assert(entry != NULL);
         assert(entry_term(entry) == i);
@@ -167,7 +167,7 @@ static void page_remove_after_test(void)
     assert(rc == RS_OK);
     assert(page_entry_count(&page) == 500);
 
-    for (int i = 0; i < 500; i++) {
+    for (uint64_t i = 0; i < 500; i++) {
         entry = page_entry_at(&page, prev_index + 1 + i);
         assert(entry != NULL);
         assert(entry_term(entry) == i);
